+++ src/errors.rs
@@ -1,4 +1,4 @@
-use std::io;
+use std::{io, str::Utf8Error};
 
 use thiserror::Error;
 
@@ -57,17 +57,26 @@ pub enum Error {
     NoSignatureFound,
 
     #[error("error creating signature: {0}")]
-    SignError(Box<dyn std::error::Error>),
+    SignError(#[source] pgp::errors::Error),
 
-    #[error("error parsing key - {details}. underlying error was: {source}")]
-    KeyLoadError {
-        source: Box<dyn std::error::Error>,
-        details: &'static str,
-    },
+    #[error("error parsing keys, failed to parse bytes as utf8 for ascii armored parsing")]
+    KeyLoadUtf8Error(
+        #[from]
+        #[source]
+        Utf8Error,
+    ),
+
+    #[error("errors parsing keys, failed to parse bytes as ascii armored key")]
+    KeyLoadSecretKeyError(
+        #[from]
+        #[source]
+        pgp::errors::Error,
+    ),
 
     #[error("error verifying signature with key {key_ref}: {source}")]
     VerificationError {
-        source: Box<dyn std::error::Error>,
+        #[source]
+        source: pgp::errors::Error,
         key_ref: String,
     },
 
@@ -110,3 +119,8 @@ impl From<TimestampError> for Error {
         Error::TimestampConv(error)
     }
 }
+
+const _: () = {
+    const fn assert_send_sync<T: Send + Sync>() {}
+    assert_send_sync::<Error>();
+};