+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+
+- `Error` now implements `Send + Sync`.
+
 ## 0.13.0
 
 ### Breaking Changes